requirements.txt
mdadiffusion>=0.0.1
PyQt5>=5.15.9
qt-material>=2.14
pyyaml>=6.0
pyinstaller
